-（迁移）批量修复 26.3 API 差异（实体/交互/粒子/方块/标签等）

+++ src/main/java/alku/csrp/entity/HiBlazeEntity.java
@@ -75,7 +75,7 @@ public final class HiBlazeEntity extends HijackedParasiteEntity implements Citad
     }
 
     @Override
-    public boolean causeFallDamage(float distance, float damageMultiplier, DamageSource source) {
+    public boolean causeFallDamage(double distance, float damageMultiplier, DamageSource source) {
         return false;
     }
